refactor(agent-hooks): merge footer hooks into one footer_guard.py
Per request: one file, two events (same pattern as security_guard.py). The
footer policy was two separate scripts (suppress_model_footer on pre_llm_call +
append_runtime_footer on on_response_end); they are one concern, so merge them
into footer_guard.py that dispatches on `event`:
  - pre_llm_call    -> inject 'don't type/imitate a footer' directive
  - on_response_end -> strip a model-typed footer at the reply end, append the
                       one true runtime model+cost footer

Behaviour is unchanged (default model+cost only, FOOTER_SHOW_TOKENS opt-in,
FOOTER_TEMPLATE / FOOTER_SUPPRESS_TEXT / FOOTER_STRIP knobs, narrow tail-only
strip). Wire both events to the same file; one approve call mounts both.

- remove suppress_model_footer.py + append_runtime_footer.py (+selftests)
- add footer_guard.py + footer_guard_selftest.py (25 cases: both events, strip
  + false-positive guards, dispatch safety)
- install_all.sh + SKILL.md updated to the single file (3 scripts / 10 entries)
